+++ src/biodwh2-gene2phenotype/src/main/java/de/unibi/agbi/biodwh2/gene2phenotype/Gene2PhenotypeDataSource.java
@@ -25,7 +25,7 @@ public class Gene2PhenotypeDataSource extends DataSource {
 
     @Override
     public DevelopmentState getDevelopmentState() {
-        return DevelopmentState.InDevelopment;
+        return DevelopmentState.Usable;
     }
 
     @Override