+++ src/biodwh2-gene2phenotype/src/main/java/de/unibi/agbi/biodwh2/gene2phenotype/model/GeneDiseasePair.java
@@ -36,7 +36,7 @@ public class GeneDiseasePair {
     @JsonProperty("prev symbols")
     public String prevSymbols;
     @JsonProperty("hgnc id")
-    public int hgncId;
+    public Integer hgncId;
     @JsonProperty("gene disease pair entry date")
     public String entryDate;
 }